@@ -71,10 +71,10 @@ class License {
         $this->schedule_hook = $this->client->slug . '_license_check_event';
 
         // Creating WP Ajax Endpoint to refresh license remotely
-        add_action( "wp_ajax_appsero_refresh_license_" . $this->client->hash, array( $this, 'refresh_license_api' ) );
+        add_action( 'wp_ajax_appsero_refresh_license_' . $this->client->hash, [ $this, 'refresh_license_api' ] );
 
         // Run hook to check license status daily
-        add_action( $this->schedule_hook, array( $this, 'check_license_status' ) );
+        add_action( $this->schedule_hook, [ $this, 'check_license_status' ] );
 
         // Active/Deactive corn schedule
         $this->run_schedule();
@@ -114,7 +114,7 @@ class License {
      * @return bool
      */
     public function check( $license_key ) {
-        $route    = 'public/license/' . $this->client->hash . '/check';
+        $route = 'public/license/' . $this->client->hash . '/check';
 
         return $this->send_request( $license_key, $route );
     }
@@ -125,7 +125,7 @@ class License {
      * @return bool
      */
     public function activate( $license_key ) {
-        $route    = 'public/license/' . $this->client->hash . '/activate';
+        $route = 'public/license/' . $this->client->hash . '/activate';
 
         return $this->send_request( $license_key, $route );
     }
@@ -136,7 +136,7 @@ class License {
      * @return bool
      */
     public function deactivate( $license_key ) {
-        $route    = 'public/license/' . $this->client->hash . '/deactivate';
+        $route = 'public/license/' . $this->client->hash . '/deactivate';
 
         return $this->send_request( $license_key, $route );
     }
@@ -144,41 +144,38 @@ class License {
     /**
      * Send common request
      *
-     * @param $license_key
-     * @param $route
-     *
      * @return array
      */
     protected function send_request( $license_key, $route ) {
-        $params = array(
+        $params = [
             'license_key' => $license_key,
             'url'         => esc_url( home_url() ),
             'is_local'    => $this->client->is_local_server(),
-        );
+        ];
 
         $response = $this->client->send_request( $params, $route, true );
 
         if ( is_wp_error( $response ) ) {
-            return array(
+            return [
                 'success' => false,
-                'error'   => $response->get_error_message()
-            );
+                'error'   => $response->get_error_message(),
+            ];
         }
 
         $response = json_decode( wp_remote_retrieve_body( $response ), true );
 
-        if ( empty( $response ) || isset( $response['exception'] )) {
-            return array(
+        if ( empty( $response ) || isset( $response['exception'] ) ) {
+            return [
                 'success' => false,
                 'error'   => $this->client->__trans( 'Unknown error occurred, Please try again.' ),
-            );
+            ];
         }
 
         if ( isset( $response['errors'] ) && isset( $response['errors']['license_key'] ) ) {
-            $response = array(
+            $response = [
                 'success' => false,
-                'error'   => $response['errors']['license_key'][0]
-            );
+                'error'   => $response['errors']['license_key'][0],
+            ];
         }
 
         return $response;
@@ -191,9 +188,9 @@ class License {
         $this->check_license_status();
 
         return wp_send_json(
-            array(
-                'message' => 'License refreshed successfully.'
-            ),
+            [
+                'message' => 'License refreshed successfully.',
+            ],
             200
         );
     }
@@ -205,8 +202,8 @@ class License {
      *
      * @return void
      */
-    public function add_settings_page( $args = array() ) {
-        $defaults = array(
+    public function add_settings_page( $args = [] ) {
+        $defaults = [
             'type'        => 'menu', // Can be: menu, options, submenu
             'page_title'  => 'Manage License',
             'menu_title'  => 'Manage License',
@@ -215,11 +212,11 @@ class License {
             'icon_url'    => '',
             'position'    => null,
             'parent_slug' => '',
-        );
+        ];
 
         $this->menu_args = wp_parse_args( $args, $defaults );
 
-        add_action( 'admin_menu', array( $this, 'admin_menu' ), 99 );
+        add_action( 'admin_menu', [ $this, 'admin_menu' ], 99 );
     }
 
     /**
@@ -247,12 +244,11 @@ class License {
      * License menu output
      */
     public function menu_output() {
-        if ( isset( $_POST['submit'] ) ) {
-            $this->license_form_submit( $_POST );
-        }
+        // process form data is submitted
+        $this->license_form_submit();
 
         $license = $this->get_license();
-        $action  = ( $license && isset( $license['status'] ) && 'activate' == $license['status'] ) ? 'deactive' : 'active';
+        $action  = ( $license && isset( $license['status'] ) && 'activate' === $license['status'] ) ? 'deactive' : 'active';
         $this->licenses_style();
         ?>
 
@@ -271,7 +267,7 @@ class License {
                     <p>
                         <?php printf( $this->client->__trans( 'Activate <strong>%s</strong> by your license key to get professional support and automatic update from your WordPress dashboard.' ), $this->client->name ); ?>
                     </p>
-                    <form method="post" action="<?php $this->form_action_url(); ?>" novalidate="novalidate" spellcheck="false">
+                    <form method="post" novalidate="novalidate" spellcheck="false">
                         <input type="hidden" name="_action" value="<?php echo $action; ?>">
                         <input type="hidden" name="_nonce" value="<?php echo wp_create_nonce( $this->client->name ); ?>">
                         <div class="license-input-fields">
@@ -281,19 +277,20 @@ class License {
                                 </svg>
                                 <input type="text" value="<?php echo $this->get_input_license_value( $action, $license ); ?>"
                                     placeholder="<?php echo esc_attr( $this->client->__trans( 'Enter your license key to activate' ) ); ?>" name="license_key"
-                                    <?php echo ( 'deactive' == $action ) ? 'readonly="readonly"' : ''; ?>
+                                    <?php echo ( 'deactive' === $action ) ? 'readonly="readonly"' : ''; ?>
                                 />
                             </div>
-                            <button type="submit" name="submit" class="<?php echo 'deactive' == $action ? 'deactive-button' : ''; ?>">
-                                <?php echo $action == 'active' ? $this->client->__trans( 'Activate License' ) : $this->client->__trans( 'Deactivate License' ); ?>
+                            <button type="submit" name="submit" class="<?php echo 'deactive' === $action ? 'deactive-button' : ''; ?>">
+                                <?php echo $action === 'active' ? $this->client->__trans( 'Activate License' ) : $this->client->__trans( 'Deactivate License' ); ?>
                             </button>
                         </div>
                     </form>
 
                     <?php
-                        if ( 'deactive' == $action && isset( $license['remaining'] ) ) {
-                            $this->show_active_license_info( $license );
-                        } ?>
+                    if ( 'deactive' === $action && isset( $license['remaining'] ) ) {
+                        $this->show_active_license_info( $license );
+                    }
+                    ?>
                 </div>
             </div> <!-- /.appsero-license-settings -->
 
@@ -305,30 +302,37 @@ class License {
     /**
      * License form submit
      */
-    public function license_form_submit( $form ) {
-        if ( ! isset( $form['_nonce'], $form['_action'] ) ) {
-            $this->error = $this->client->__trans( 'Please add all information' );
+    public function license_form_submit() {
+        if ( ! isset( $_POST['_nonce'] ) ) {
+            return;
+        }
+
+        if ( ! wp_verify_nonce( sanitize_key( wp_unslash( $_POST['_nonce'] ) ), $this->client->name ) ) {
+            $this->error = $this->client->__trans( 'Nonce vefification failed.' );
 
             return;
         }
 
-        if ( ! wp_verify_nonce( $form['_nonce'], $this->client->name ) ) {
-            $this->error = $this->client->__trans( "You don't have permission to manage license." );
+        if ( ! current_user_can( 'manage_options' ) ) {
+            $this->error = $this->client->__trans( 'You don\'t have permission to manage license.' );
 
             return;
         }
 
-        switch ( $form['_action'] ) {
+        $license_key = isset( $_POST['license_key'] ) ? sanitize_text_field( wp_unslash( $_POST['license_key'] ) ) : '';
+        $action      = isset( $_POST['_action'] ) ? sanitize_text_field( wp_unslash( $_POST['_action'] ) ) : '';
+
+        switch ( $action ) {
             case 'active':
-                $this->active_client_license( $form );
+                $this->active_client_license( $license_key );
                 break;
 
             case 'deactive':
-                $this->deactive_client_license( $form );
+                $this->deactive_client_license();
                 break;
 
             case 'refresh':
-                $this->refresh_client_license( $form );
+                $this->refresh_client_license();
                 break;
         }
     }
@@ -369,7 +373,7 @@ class License {
 
         $license = $this->get_license();
 
-        if ( ! empty( $license['key'] ) && isset( $license['status'] ) && $license['status'] == 'activate' ) {
+        if ( ! empty( $license['key'] ) && isset( $license['status'] ) && $license['status'] === 'activate' ) {
             $this->is_valid_license = true;
         } else {
             $this->is_valid_license = false;
@@ -384,8 +388,8 @@ class License {
     public function is_valid_by( $option, $value ) {
         $license = $this->get_license();
 
-        if ( ! empty( $license['key'] ) && isset( $license['status'] ) && $license['status'] == 'activate' ) {
-            if ( isset( $license[ $option ] ) && $license[ $option ] == $value ) {
+        if ( ! empty( $license['key'] ) && isset( $license['status'] ) && $license['status'] === 'activate' ) {
+            if ( isset( $license[ $option ] ) && $license[ $option ] === $value ) {
                 return true;
             }
         }
@@ -548,12 +552,13 @@ class License {
             <div class="single-license-info">
                 <h3><?php $this->client->_etrans( 'Expires in' ); ?></h3>
                 <?php
-                    if ( false !== $license['expiry_days'] ) {
-                        $occupied = $license['expiry_days'] > 21 ? '' : 'occupied';
-                        echo '<p class="' . $occupied . '">' . $license['expiry_days'] . ' days</p>';
-                    } else {
-                        echo '<p>' . $this->client->__trans( 'Never' ) . '</p>';
-                    } ?>
+                if ( false !== $license['expiry_days'] ) {
+                    $occupied = $license['expiry_days'] > 21 ? '' : 'occupied';
+                    echo '<p class="' . $occupied . '">' . $license['expiry_days'] . ' days</p>';
+                } else {
+                    echo '<p>' . $this->client->__trans( 'Never' ) . '</p>';
+                }
+                ?>
             </div>
         </div>
         <?php
@@ -568,7 +573,7 @@ class License {
             <div class="notice notice-error is-dismissible appsero-license-section">
                 <p><?php echo $this->error; ?></p>
             </div>
-        <?php
+			<?php
         }
 
         if ( ! empty( $this->success ) ) {
@@ -576,7 +581,7 @@ class License {
             <div class="notice notice-success is-dismissible appsero-license-section">
                 <p><?php echo $this->success; ?></p>
             </div>
-        <?php
+			<?php
         }
         echo '<br />';
     }
@@ -594,16 +599,16 @@ class License {
             </svg>
             <span><?php echo $this->client->__trans( 'Activate License' ); ?></span>
 
-            <?php if ( $license && $license['key'] ) : ?>
-            <form method="post" class="appsero-license-right-form" action="<?php $this->form_action_url(); ?>" novalidate="novalidate" spellcheck="false">
+            <?php if ( $license && $license['key'] ) { ?>
+            <form method="post" class="appsero-license-right-form" novalidate="novalidate" spellcheck="false">
                 <input type="hidden" name="_action" value="refresh">
-                <input type="hidden" name="_nonce" value="<?php echo wp_create_nonce( $this->client->name ); ?>">
+                <input type="hidden" name="_appsero_license_nonce" value="<?php echo wp_create_nonce( $this->client->name ); ?>">
                 <button type="submit" name="submit" class="appsero-license-refresh-button">
                     <span class="dashicons dashicons-update"></span>
                     <?php echo $this->client->__trans( 'Refresh License' ); ?>
                 </button>
             </form>
-            <?php endif; ?>
+            <?php } ?>
 
         </div>
         <?php
@@ -612,15 +617,14 @@ class License {
     /**
      * Active client license
      */
-    private function active_client_license( $form ) {
-        if ( empty( $form['license_key'] ) ) {
+    private function active_client_license( $license_key ) {
+        if ( empty( $license_key ) ) {
             $this->error = $this->client->__trans( 'The license key field is required.' );
 
             return;
         }
 
-        $license_key = sanitize_text_field( $form['license_key'] );
-        $response    = $this->activate( $license_key );
+        $response = $this->activate( $license_key );
 
         if ( ! $response['success'] ) {
             $this->error = $response['error'] ? $response['error'] : $this->client->__trans( 'Unknown error occurred.' );
@@ -628,7 +632,7 @@ class License {
             return;
         }
 
-        $data = array(
+        $data = [
             'key'              => $license_key,
             'status'           => 'activate',
             'remaining'        => $response['remaining'],
@@ -637,7 +641,7 @@ class License {
             'title'            => $response['title'],
             'source_id'        => $response['source_identifier'],
             'recurring'        => $response['recurring'],
-        );
+        ];
 
         update_option( $this->option_key, $data, false );
 
@@ -647,7 +651,7 @@ class License {
     /**
      * Deactive client license
      */
-    private function deactive_client_license( $form ) {
+    private function deactive_client_license() {
         $license = $this->get_license();
 
         if ( empty( $license['key'] ) ) {
@@ -658,10 +662,10 @@ class License {
 
         $response = $this->deactivate( $license['key'] );
 
-        $data = array(
+        $data = [
             'key'    => '',
             'status' => 'deactivate',
-        );
+        ];
 
         update_option( $this->option_key, $data, false );
 
@@ -677,11 +681,12 @@ class License {
     /**
      * Refresh Client License
      */
-    private function refresh_client_license( $form = null ) {
+    private function refresh_client_license() {
         $license = $this->get_license();
 
-        if( !$license || ! isset( $license['key'] ) || empty( $license['key'] ) ) {
-            $this->error = $this->client->__trans( "License key not found" );
+        if ( ! $license || ! isset( $license['key'] ) || empty( $license['key'] ) ) {
+            $this->error = $this->client->__trans( 'License key not found' );
+
             return;
         }
 
@@ -695,12 +700,12 @@ class License {
      */
     private function create_menu_page() {
         call_user_func(
-            'add_' . 'menu' . '_page',
+            'add_menu_page',
             $this->menu_args['page_title'],
             $this->menu_args['menu_title'],
             $this->menu_args['capability'],
             $this->menu_args['menu_slug'],
-            array( $this, 'menu_output' ),
+            [ $this, 'menu_output' ],
             $this->menu_args['icon_url'],
             $this->menu_args['position']
         );
@@ -711,13 +716,13 @@ class License {
      */
     private function create_submenu_page() {
         call_user_func(
-            'add_' . 'submenu' . '_page',
+            'add_submenu_page',
             $this->menu_args['parent_slug'],
             $this->menu_args['page_title'],
             $this->menu_args['menu_title'],
             $this->menu_args['capability'],
             $this->menu_args['menu_slug'],
-            array( $this, 'menu_output' ),
+            [ $this, 'menu_output' ],
             $this->menu_args['position']
         );
     }
@@ -727,12 +732,12 @@ class License {
      */
     private function create_options_page() {
         call_user_func(
-            'add_' . 'options' . '_page',
+            'add_options_page',
             $this->menu_args['page_title'],
             $this->menu_args['menu_title'],
             $this->menu_args['capability'],
             $this->menu_args['menu_slug'],
-            array( $this, 'menu_output' ),
+            [ $this, 'menu_output' ],
             $this->menu_args['position']
         );
     }
@@ -761,46 +766,34 @@ class License {
     private function run_schedule() {
         switch ( $this->client->type ) {
             case 'plugin':
-                register_activation_hook( $this->client->file, array( $this, 'schedule_cron_event' ) );
-                register_deactivation_hook( $this->client->file, array( $this, 'clear_scheduler' ) );
+                register_activation_hook( $this->client->file, [ $this, 'schedule_cron_event' ] );
+                register_deactivation_hook( $this->client->file, [ $this, 'clear_scheduler' ] );
                 break;
 
             case 'theme':
-                add_action( 'after_switch_theme', array( $this, 'schedule_cron_event' ) );
-                add_action( 'switch_theme', array( $this, 'clear_scheduler' ) );
+                add_action( 'after_switch_theme', [ $this, 'schedule_cron_event' ] );
+                add_action( 'switch_theme', [ $this, 'clear_scheduler' ] );
                 break;
         }
-    }
-
-    /**
-     * Form action URL
-     */
-    private function form_action_url() {
-        $url = add_query_arg(
-            $_GET,
-            admin_url( basename( $_SERVER['SCRIPT_NAME'] ) )
-        );
-
-        echo apply_filters( 'appsero_client_license_form_action', $url );
     }
 
     /**
      * Get input license key
      *
-     * @param  $action
-     *
      * @return $license
      */
     private function get_input_license_value( $action, $license ) {
-        if ( 'active' == $action ) {
+        if ( 'active' === $action ) {
             return isset( $license['key'] ) ? $license['key'] : '';
         }
 
-        if ( 'deactive' == $action ) {
+        if ( 'deactive' === $action ) {
             $key_length = strlen( $license['key'] );
 
             return str_pad(
-                substr( $license['key'], 0, $key_length / 2 ), $key_length, '*'
+                substr( $license['key'], 0, $key_length / 2 ),
+                $key_length,
+                '*'
             );
         }
 
