@@ -1,5 +1,8 @@
 <?php
+
 namespace Appsero;
+
+use stdClass;
 
 /**
  * Appsero Updater
@@ -16,19 +19,25 @@ class Updater {
     protected $client;
 
     /**
+     * Cache key
+     *
+     * @var string
+     */
+    protected $cache_key;
+
+    /**
      * Initialize the class
      *
      * @param Appsero\Client
      */
     public function __construct( Client $client ) {
-
         $this->client    = $client;
         $this->cache_key = 'appsero_' . md5( $this->client->slug ) . '_version_info';
 
         // Run hooks.
-        if ( $this->client->type == 'plugin' ) {
+        if ( $this->client->type === 'plugin' ) {
             $this->run_plugin_hooks();
-        } elseif ( $this->client->type == 'theme' ) {
+        } elseif ( $this->client->type === 'theme' ) {
             $this->run_theme_hooks();
         }
     }
@@ -39,8 +48,8 @@ class Updater {
      * @return void
      */
     public function run_plugin_hooks() {
-        add_filter( 'pre_set_site_transient_update_plugins', array( $this, 'check_plugin_update' ) );
-        add_filter( 'plugins_api', array( $this, 'plugins_api_filter' ), 10, 3 );
+        add_filter( 'pre_set_site_transient_update_plugins', [ $this, 'check_plugin_update' ] );
+        add_filter( 'plugins_api', [ $this, 'plugins_api_filter' ], 10, 3 );
     }
 
     /**
@@ -49,7 +58,7 @@ class Updater {
      * @return void
      */
     public function run_theme_hooks() {
-        add_filter( 'pre_set_site_transient_update_themes', array( $this, 'check_theme_update' ) );
+        add_filter( 'pre_set_site_transient_update_themes', [ $this, 'check_theme_update' ] );
     }
 
     /**
@@ -59,10 +68,10 @@ class Updater {
         global $pagenow;
 
         if ( ! is_object( $transient_data ) ) {
-            $transient_data = new \stdClass;
+            $transient_data = new stdClass();
         }
 
-        if ( 'plugins.php' == $pagenow && is_multisite() ) {
+        if ( 'plugins.php' === $pagenow && is_multisite() ) {
             return $transient_data;
         }
 
@@ -73,7 +82,6 @@ class Updater {
         $version_info = $this->get_version_info();
 
         if ( false !== $version_info && is_object( $version_info ) && isset( $version_info->new_version ) ) {
-
             unset( $version_info->sections );
 
             // If new version available then set to `response`
@@ -84,7 +92,7 @@ class Updater {
                 $transient_data->no_update[ $this->client->basename ] = $version_info;
             }
 
-            $transient_data->last_checked = time();
+            $transient_data->last_checked                       = time();
             $transient_data->checked[ $this->client->basename ] = $this->client->project_version;
         }
 
@@ -94,19 +102,19 @@ class Updater {
     /**
      * Get version info from database
      *
-     * @return Object or Boolean
+     * @return object or Boolean
      */
     private function get_cached_version_info() {
         global $pagenow;
 
         // If updater page then fetch from API now
-        if ( 'update-core.php' == $pagenow ) {
+        if ( 'update-core.php' === $pagenow ) {
             return false; // Force to fetch data
         }
 
         $value = get_transient( $this->cache_key );
 
-        if( ! $value && ! isset( $value->name ) ) {
+        if ( ! $value && ! isset( $value->name ) ) {
             return false; // Cache is expired
         }
 
@@ -142,16 +150,15 @@ class Updater {
      * Get plugin info from Appsero
      */
     private function get_project_latest_version() {
-
         $license = $this->client->license()->get_license();
 
-        $params = array(
+        $params = [
             'version'     => $this->client->project_version,
             'name'        => $this->client->name,
             'slug'        => $this->client->slug,
             'basename'    => $this->client->basename,
             'license_key' => ! empty( $license ) && isset( $license['key'] ) ? $license['key'] : '',
-        );
+        ];
 
         $route = 'update/' . $this->client->hash . '/check';
 
@@ -185,19 +192,18 @@ class Updater {
     /**
      * Updates information on the "View version x.x details" page with custom data.
      *
-     * @param mixed   $data
-     * @param string  $action
-     * @param object  $args
+     * @param mixed  $data
+     * @param string $action
+     * @param object $args
      *
      * @return object $data
      */
     public function plugins_api_filter( $data, $action = '', $args = null ) {
-
-        if ( $action != 'plugin_information' ) {
+        if ( $action !== 'plugin_information' ) {
             return $data;
         }
 
-        if ( ! isset( $args->slug ) || ( $args->slug != $this->client->slug ) ) {
+        if ( ! isset( $args->slug ) || ( $args->slug !== $this->client->slug ) ) {
             return $data;
         }
 
@@ -211,10 +217,10 @@ class Updater {
         global $pagenow;
 
         if ( ! is_object( $transient_data ) ) {
-            $transient_data = new \stdClass;
+            $transient_data = new stdClass();
         }
 
-        if ( 'themes.php' == $pagenow && is_multisite() ) {
+        if ( 'themes.php' === $pagenow && is_multisite() ) {
             return $transient_data;
         }
 
@@ -225,7 +231,6 @@ class Updater {
         $version_info = $this->get_version_info();
 
         if ( false !== $version_info && is_object( $version_info ) && isset( $version_info->new_version ) ) {
-
             // If new version available then set to `response`
             if ( version_compare( $this->client->project_version, $version_info->new_version, '<' ) ) {
                 $transient_data->response[ $this->client->slug ] = (array) $version_info;
@@ -234,7 +239,7 @@ class Updater {
                 $transient_data->no_update[ $this->client->slug ] = (array) $version_info;
             }
 
-            $transient_data->last_checked = time();
+            $transient_data->last_checked                   = time();
             $transient_data->checked[ $this->client->slug ] = $this->client->project_version;
         }
 
@@ -254,5 +259,4 @@ class Updater {
 
         return $version_info;
     }
-
 }
